Fix server starting
Testing pulling the commit and launching using nginx start revealed that
nginx failed to launch because it needed a folder called temp created.
Github ignored folders by this name by default. Changed the bat file to
create the temp directory on launch. Added ignore for the .pid file as
it simply tracks a running process and does not need to be synced.