Stop scrolling if mouse leaves canvas
the moveMouse event is only captured on the canvas. Since the player bar
is no longer part of the canvas object, moveMouse events are not sent
when the mouse is over that object, and so the previous mouse check was
failing. Fixed by capturing the mouseOut event.

Fixes #53 by capturing mouse out event.